+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/beneficiary/BeneficiaryList.kt
@@ -88,6 +88,10 @@ fun BeneficiaryListScreen(
         ?.savedStateHandle
         ?.get<String>("success_message")
 
+    LaunchedEffect(Unit) {
+        viewModel.getBeneficiaries()
+    }
+
     LaunchedEffect(successMessage) {
         successMessage?.let { message ->
             scope.launch {