+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/workshop/ModifyWorkshopViewModel.kt
@@ -112,7 +112,7 @@ class ModifyWorkshopViewModel @Inject constructor(
                 val workshopData = repository.getWorkshopsById(idWorkshop)
                 _workshop.value = workshopData
 
-                Log.d("WORKSHOP_DEBUG", "Hora Inicio (Cruda): ${workshopData?.startHour}")
+                Log.d("WORKSHOP_DEBUG", "Workshop: ${workshopData}")
                 if (workshopData != null) {
                     val cleanDate = formatWorkshopDate(workshopData.date)
                     _formData.update {
@@ -128,7 +128,6 @@ class ModifyWorkshopViewModel @Inject constructor(
                         )
                     }
                     _formattedDate.value = formatWorkshopDate(workshopData.date)
-                    Log.d("WORKSHOP_DEBUG", "  startHour: ${_formData.value.startHour}")
                 } else {
                     _workLoadError.value = "No se encontró el taller"
                 }
@@ -148,12 +147,13 @@ class ModifyWorkshopViewModel @Inject constructor(
             val instant = Instant.parse(dateString)
             val zonedDateTime = instant.atZone(ZoneId.systemDefault())
 
-            val formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd")
+            val formatter = DateTimeFormatter.ofPattern("dd/MM/yyyy")
             zonedDateTime.format(formatter)
         } catch (e: Exception) {
 
             if (dateString.matches(Regex("^\\d{4}-\\d{2}-\\d{2}"))) {
-                dateString
+                val parts = dateString.split("-")
+                "${parts[2]}/${parts[1]}/${parts[0]}"
             } else {
                 Log.e("WORKSHOP_DEBUG", "Error al formatear o validar fecha: $dateString", e)
                 ""
@@ -178,6 +178,7 @@ class ModifyWorkshopViewModel @Inject constructor(
                 image = _formData.value.image,
                 videoTraining = _formData.value.videoTraining
             )
+            Log.d("WORKSHOP_DEBUG", "Fecha: ${_formData?.value?.date}")
             postModifyWorkshop(workshopDto).collect { result ->
                 _uiState.update { state ->
                     when (result) {
@@ -226,7 +227,7 @@ class ModifyWorkshopViewModel @Inject constructor(
         val data = _formData.value
         val errors = mutableMapOf<String, Boolean>()
         val hourRegex = Regex("^([01]?\\d|2[0-3]):[0-5]\\d$")
-        val dateRegex = Regex("^\\d{4}-\\d{2}-\\d{2}$")
+        val dateRegex = Regex("^\\d{2}/\\d{2}/\\d{4}$")
 
         if (data.name.isBlank()) errors["name"] = true
         if (data.startHour.isBlank()) errors["startHour"] = true