+++ CHANGELOG.md
@@ -4,13 +4,24 @@
 - Functional go back button to Supplies List Screen
 - Functional go back buttons to Orders and Inventory Screens
 
-### Fixed
-- Fix beneficiary list navbar
-- Fix the go back buttons of Products, Product Batches and Supplies to be the same
-- - Fix the go back button of teh Beneficiary Screen to be the same icon of the Workshop screen
-
 ### Deleted
 - All the NotificationIcon buttons of the screens
+- Scroll for login, home and inventory screens.
+
+### Fixed
+- Fix beneficiary list navbar
+- Fix workshops date input
+- Fix workshops hour input
+- Fix Workshops navbar
+- Fix beneficiary registration navBar appearing
+- Fix view after deleting a beneficiary
+- Fix date input
+- Fix scroll bar for supply batch views
+- Input Validations
+- Fix view to not show the navbar in beneficiary details
+- Fix beneficiary list navbar
+- Fix the go back buttons of Products, Product Batches and Supplies to be the same
+- Fix the go back button of teh Beneficiary Screen to be the same icon of the Workshop screen
 
 ## [0.1.0] - 19/11/2025
 ### Added